head-count/requirements.txt
flask==3.0.0
pillow==10.1.0
numpy==1.24.3
# CPU-only PyTorch
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.230
opencv-python-headless==4.8.1.78
segmentation-models-pytorch
